Move QuestionAlertsMenuItem
